Move Alphanum comparator into own package, add license file
